+++ gears/compressors/cssmin.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 try:
     from cssmin import cssmin
     cssmin_available = True