+++ gears/environment.py
@@ -1,11 +1,17 @@
+import gzip
 import os
-import sys
 
 from .asset_attributes import AssetAttributes
 from .assets import build_asset
 from .cache import SimpleCache
+from .compat import bytes
 from .exceptions import FileNotFound
-from .processors import DirectivesProcessor
+from .manifest import Manifest
+from .processors import (
+    DirectivesProcessor,
+    HexdigestPathsProcessor,
+    SemicolonsProcessor
+)
 from .utils import get_condition_func
 
 
@@ -125,6 +131,10 @@ class Postprocessors(Processors):
     used in the order they were added.
     """
 
+    def register_defaults(self):
+        self.register('application/javascript', SemicolonsProcessor.as_handler())
+        self.register('text/css', HexdigestPathsProcessor.as_handler())
+
 
 class Compressors(dict):
     """The registry for asset compressors. It acts like a dictionary with
@@ -212,10 +222,23 @@ class Environment(object):
                   store compilation results.
     """
 
-    def __init__(self, root, public_assets=DEFAULT_PUBLIC_ASSETS, cache=None):
+    def __init__(self, root, public_assets=DEFAULT_PUBLIC_ASSETS,
+                 manifest_path=None, cache=None, gzip=False):
         self.root = root
         self.public_assets = [get_condition_func(c) for c in public_assets]
-        self.cache = cache if cache is not None else SimpleCache()
+
+        if manifest_path is not None:
+            self.manifest_path = manifest_path
+        else:
+            self.manifest_path = os.path.join(self.root, '.manifest.json')
+        self.manifest = Manifest(self.manifest_path)
+
+        if cache is not None:
+            self.cache = cache
+        else:
+            self.cache = SimpleCache()
+
+        self.gzip = gzip
 
         #: The registry for file finders. See
         #: :class:`~gears.environment.Finders` for more information.
@@ -260,6 +283,7 @@ class Environment(object):
         """Register default compilers, preprocessors and MIME types."""
         self.mimetypes.register_defaults()
         self.preprocessors.register_defaults()
+        self.postprocessors.register_defaults()
 
     def find(self, item, logical=False):
         """Find files using :attr:`finders` registry. The ``item`` parameter
@@ -341,13 +365,13 @@ class Environment(object):
             logical_path = os.path.normpath(asset_attributes.logical_path)
             if self.is_public(logical_path):
                 asset = build_asset(self, logical_path)
-                if sys.version_info < (3, 0):
-                    source = str(asset)
-                else:
-                    source = bytes(asset)
-                self.save_file(logical_path, source)
+                source = bytes(asset)
+                self.save_file(logical_path, source, asset.gzippable)
+                self.save_file(asset.hexdigest_path, source, asset.gzippable)
+                self.manifest.files[logical_path] = asset.hexdigest_path
+        self.manifest.dump()
 
-    def save_file(self, path, source):
+    def save_file(self, path, source, gzippable=False):
         filename = os.path.join(self.root, path)
         path = os.path.dirname(filename)
         if not os.path.exists(path):
@@ -356,6 +380,9 @@ class Environment(object):
             raise OSError("%s exists and is not a directory." % path)
         with open(filename, 'wb') as f:
             f.write(source)
+        if self.gzip and gzippable:
+            with gzip.open('{}.gz'.format(filename), 'wb') as f:
+                f.write(source)
 
     def is_public(self, logical_path):
         return any(condition(logical_path) for condition in self.public_assets)