+++ tests/test_assets.py
@@ -1,50 +1,103 @@
-import codecs
-import os
-import sys
+from gears.assets import (
+    CircularDependencyError, BaseAsset, Asset, StaticAsset, build_asset,
+    strip_fingerprint
+)
+from gears.compat import str, bytes
 
-from gears.assets import CircularDependencyError, Asset
-from gears.environment import Environment
-from gears.finders import FileSystemFinder
-
+from mock import sentinel, Mock
 from unittest2 import TestCase
+from .helpers import GearsTestCase
 
 
-TESTS_DIR = os.path.dirname(__file__)
-FIXTURES_DIR = os.path.join(TESTS_DIR, 'fixtures', 'assets')
+class AssetTests(GearsTestCase):
 
-
-def read(file):
-    with codecs.open(file, encoding='utf-8') as f:
-        return f.read()
-
-
-class AssetTests(TestCase):
-
-    def get_fixture_path(self, fixture):
-        return os.path.join(FIXTURES_DIR, fixture)
-
-    def get_finder(self, fixture):
-        return FileSystemFinder([self.get_fixture_path(fixture)])
-
-    def get_environment(self, fixture):
-        environment = Environment(os.path.join(TESTS_DIR, 'static'))
-        environment.finders.register(self.get_finder(fixture))
-        environment.mimetypes.register_defaults()
-        environment.preprocessors.register_defaults()
-        return environment
-
-    def get_asset(self, fixture):
-        return Asset(*self.get_environment(fixture).find('source.js'))
+    fixtures_root = 'assets'
 
     def test_circular_dependency(self):
         with self.assertRaises(CircularDependencyError):
             asset = self.get_asset('circular_dependency')
 
     def test_unicode_support(self):
-        output = read(os.path.join(FIXTURES_DIR, 'unicode_support', 'output.js'))
         asset = self.get_asset('unicode_support')
-        if sys.version_info < (3, 0):
-            asset_output = unicode(asset)
-        else:
-            asset_output = str(asset)
-        self.assertEqual(asset_output, output)
+        output = self.get_output('unicode_support')
+        self.assertEqual(str(asset), output)
+
+    def test_is_iterable(self):
+        asset = self.get_asset('unicode_support')
+        tuple(asset)
+
+    def test_is_convertible_to_bytes(self):
+        asset = self.get_asset('unicode_support')
+        bytes(asset)
+
+
+class StaticAssetTests(GearsTestCase):
+
+    fixtures_root = 'assets'
+
+    def test_source(self):
+        asset = self.get_static_asset('static_source')
+        asset.source
+
+    def test_hexdigest(self):
+        asset = self.get_static_asset('static_source')
+        self.assertEqual(
+            asset.hexdigest,
+            'c8a756475599e6e3c904b24077b4b0a31983752c',
+        )
+
+    def test_is_iterable(self):
+        asset = self.get_static_asset('static_source')
+        tuple(asset)
+
+    def test_is_convertible_to_bytes(self):
+        asset = self.get_static_asset('static_source')
+        bytes(asset)
+
+
+class HexdigestPathTests(TestCase):
+
+    def get_asset(self, logical_path):
+        attributes = Mock(logical_path=logical_path)
+        asset = BaseAsset(attributes, sentinel.absolute_path)
+        asset.final_hexdigest = '123456'
+        return asset
+
+    def test_hexdigest_path(self):
+        def check(logical_path, result):
+            asset = self.get_asset(logical_path)
+            self.assertEqual(asset.hexdigest_path, result)
+
+        check('css/style.css', 'css/style.123456.css')
+        check('css/style.min.css', 'css/style.min.123456.css')
+
+
+class BuildAssetTests(GearsTestCase):
+
+    fixtures_root = 'assets'
+
+    def setUp(self):
+        self.environment = self.get_environment('build_asset')
+
+    def test_has_processors(self):
+        asset = build_asset(self.environment, 'source.js')
+        self.assertIsInstance(asset, Asset)
+
+    def test_has_no_processors(self):
+        asset = build_asset(self.environment, 'source.md')
+        self.assertIsInstance(asset, StaticAsset)
+
+    def test_strips_fingerprint(self):
+        path = 'source.38976434f000bf447d2dc6980894986aaf4e82d0.js'
+        asset = build_asset(self.environment, path)
+        self.assertEqual(asset.attributes.logical_path, 'source.js')
+
+
+class StripFingerprintTests(TestCase):
+
+    def test_strips_fingerprint(self):
+        path = 'source.38976434f000bf447d2dc6980894986aaf4e82d0.js'
+        self.assertEqual(strip_fingerprint(path), 'source.js')
+
+    def test_skips_paths_without_fingerprint(self):
+        self.assertEqual(strip_fingerprint('source.js'), 'source.js')